move Config to extension so others may import it
Can't live in the main package if it'll be used by other packages.
