+++ extension/config.go
@@ -1,4 +1,4 @@
-package main
+package extension
 
 import (
 	"os"
