+++ extension/config_test.go
@@ -1,4 +1,4 @@
-package main
+package extension
 
 import (
 	"testing"
